didn't want my email posted in the middle of nyc
which is strange because it's already on the **public** internet. idk i
put up the first flyer and immediately felt like i shouldn't put it out
there like that. i tried to put them up in courant (that feels safer?),
but you now have to tap an ID to get in, and not just flash it to the
security guard :'(. I even tried to tell them i was an alum and really
had to pee, but they said something about covid protocols. wack.

so now i'll have to go back and put up these new ones (and yes i made a
new email just for this)

unfortunately I don't think I can hide my name in the google sheet if
i'm in the document the same time someone else is. don't want to have to
create a new google account because i'm pretty sure they ask for a phone
number for that shit